Initial refactor commit
