--- .abstruse.yml
@@ -1,24 +1,0 @@
-image: ubuntu:focal
-
-before_install:
-  - apt update && apt install -y wget
-  - wget -qO- https://raw.githubusercontent.com/nvm-sh/nvm/v0.37.2/install.sh | bash
-  - nvm install 16
-
-install:
-  - npm install
-
-script:
-  - if [[ "$SCRIPT" ]]; then npm run $SCRIPT; fi
-
-cache:
-  - node_modules
-
-before_deploy:
-  - apt update && apt install -y wget
-  - wget -qO- https://raw.githubusercontent.com/nvm-sh/nvm/v0.37.2/install.sh | bash
-  - nvm install 16
-  - npm install
-
-deploy:
-  - node index.js
