+++ commands/info/help.js
@@ -2,12 +2,12 @@
 const Discord = require('discord.js');
 const fs = require("fs");
 var read = require('fs-readdir-recursive')
-const commandFiles = read('./commands');//fs.readdirSync('./commands');//.filter(file => file.endsWith('.js'));
+const commandFiles = read('@commands');//fs.readdirSync('./commands');//.filter(file => file.endsWith('.js'));
 const owner = process.env.OWNER;
 const prefix = process.env.BOTPREFIX;
 const { t } = require('localizify');
-const { sendTimedMessage } = require("../../utils/messages.js");
-const { tn } = require('../../utils/utils.js');
+const { sendTimedMessage } = require("@modules/utils/messages.js");
+const { tn } = require('@modules/utils/utils.js');
 
 // export module
 module.exports = {
