+++ commands/info/userinfo.js
@@ -6,7 +6,7 @@ const fs = require('fs');
 var {image_search} = require('duckduckgo-images-api');
 const axios = require('axios');
 var remote = require('remote-file-size')
-const { getBotVersion, getFooter, getTulipoSettings, setTulipoSettings } = require("../../utils/data.js");
+const { getBotVersion, getFooter, getTulipoSettings, setTulipoSettings } = require("@modules/utils/data.js");
 
 // export module
 module.exports = {
