+++ config/bot.js
@@ -1,4 +1,6 @@
 module.exports = {
+    name: 'Zumito',
+    shortUrl: 'tuipo.ga',
     embeds: {
         color: "#f982a2"
     }
