+++ index.js
@@ -1,3 +1,5 @@
+require('module-alias/register');
+
 // Load .env variables
 require("dotenv").config();
