+++ jsconfig.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "target": "es2020",
+        "jsx": "preserve",
+        "strictFunctionTypes": true,
+        "baseUrl": "./",
+        "paths": {
+            "@modules/*": ["modules/*"],
+            "@events/*": ["events/*"],
+            "@commands/*": ["commands/*"],
+            "@config/*": ["config/*"],
+        }
+    },
+    "exclude": [
+        "node_modules",
+        "**/node_modules/*"
+    ]
+}
